+++ lib/fp_util.dart
@@ -4,6 +4,7 @@ export 'src/constants/constants.dart';
 export 'src/dialogs/dialogs.dart';
 export 'src/extensions/extensions.dart';
 export 'src/form/form.dart';
+export 'src/helpers/clipboard_helper.dart';
 export 'src/helpers/image_picker_helper.dart';
 export 'src/input_formatters/input_formatters.dart';
 export 'src/logger/logger.dart';